containers/shiva/requirements.txt
hazelcast-python-client
cherrypy
pymongo
jinja2
